// src/app/models/phoneInfo.ts
export interface PhoneInfo{
  createAt: null;
  lastModifiedAt: null;
  createdBy: null;
  lastModifiedBy: null;
  serialVersionUID: number;
  serviceNo: string;
  locationCode: string;
  crmStatus: number;
  crmLastModified: number;
  rtsId: null;
  assignedId: number;
  blockId: number;
  providerId: number;
  productId: null;
  modifiedBy: string;
  lastModified: number;
  unmatchedId: null;
}